Logic to recover from out of range lower limit errors of the API
Sometimes when making the request, due timezone transformations it is possible to set the beginning of the query before the lower limit of allowed range. In this case FMIApi will return special
exception with info about allowed lower limit. This commit code will try to parse given info and retry once with given lower limit.